Split `tonic::api::points_common`...
...into `tonic::api::query_common` and `tonic::api::update_common`
